requirements.txt
streamlit
opencv-python-headless
scikit-learn
scikit-image
joblib
numpy
pillow 